Finished all functionalities and UI design as well, just need to do final testing then done

+++ client/src/components/knowledgeHub/MyArticles.js
@@ -15,6 +15,8 @@ import { formatToJustDate, formatDate } from '../../utils';
 import { Link } from 'react-router-dom';
 import ROUTES from '../../routes';
 import { getRoleText } from '../../utils';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faUser, faTimesCircle } from '@fortawesome/free-solid-svg-icons';
 
 
 const MyArticles = () => {
@@ -67,6 +69,7 @@ const MyArticles = () => {
 
     
     const handleCreateMyArticle = async (e) => {
+
         e.preventDefault();
         try {
             //console.log("BEFORE SUBMITTING TO DB: ", newArticle, ", ", user);
@@ -83,12 +86,14 @@ const MyArticles = () => {
     };
 
     const handleDeleteMyArticle = async (article_id) => {
-        try {
-            await articlesAPI.deleteArticle(article_id); // Deleting article
-            // Don't need to specifically delete comments as cascade effect does it for us automatically
-            dispatch(deleteMyArticle({id: article_id}));
-        } catch (error) {
-            console.error('Error deleting article:', error);
+        if (window.confirm('Are you sure you want to delete this article?')) {
+            try {
+                await articlesAPI.deleteArticle(article_id); // Deleting article
+                // Don't need to specifically delete comments as cascade effect does it for us automatically
+                dispatch(deleteMyArticle({id: article_id}));
+            } catch (error) {
+                console.error('Error deleting article:', error);
+            }
         }
     };
 
@@ -113,40 +118,45 @@ const MyArticles = () => {
 
     return (
         <div className='my-articles-container'>
-            <h1>My Articles</h1>
-            <button onClick={checkIfUserCanCreateArticle}>+</button>
+            <div className='top-section-of-my-articles'>
+                <h1>My Articles</h1>
+                <button className='create-new-article-user-button' onClick={checkIfUserCanCreateArticle}>+</button>
+            </div>
+            
 
             {showCreateForm && (
-                <form onSubmit={handleCreateMyArticle}>
-                    <button onClick={() => setShowCreateForm(false)}>X</button>
+                <form className="my-form" onSubmit={handleCreateMyArticle}>
+                    <button className="close-button" onClick={() => setShowCreateForm(false)}>X</button>
                     <input
+                        className="form-input-title"
                         type="text"
                         placeholder="Title"
                         value={newArticle.title}
                         onChange={(e) => setNewArticle({ ...newArticle, title: e.target.value })}
                     />
                     <textarea
+                        className="form-input-content"
                         placeholder="Content"
                         value={newArticle.content}
                         onChange={(e) => setNewArticle({ ...newArticle, content: e.target.value })}
                     />
-                    <button type="submit">Submit</button>
+                    <button className="submit-button" type="submit">Submit</button>
                 </form>
             )}
 
 
-            {fetchingUsersArticles && <p>Loading...</p>}
+            {fetchingUsersArticles && <p className='loading-text-my-articles'>Loading...</p>}
             <ul className='my-articles-list-container'>
                 {Object.values(usersArticles).map((article) => (
                     <div className='my-article-container'>
-                        <Link to={ROUTES.article(article.id, 'mine')}>
+                        <Link className='my-article-link' to={ROUTES.article(article.id, 'mine')}>
                             <li key={article.id}>
-                                <h2>{article.title}</h2>
+                                <p>{article.title}</p>
                                 <p>{article.author_username} ({getRoleText(article.author_role)})</p>
                                 <p>{article.created_at}</p>                
                             </li>
                         </Link>
-                        <button onClick={() => handleDeleteMyArticle(article.id)}>Delete</button>
+                        <FontAwesomeIcon icon={faTimesCircle} className="delete-my-article-button" onClick={() => handleDeleteMyArticle(article.id)} />
                     </div>
                 ))}
             </ul>
